Addressed colden's PR feedback.
