+++ lib/aws_mqtt.ts
@@ -1,0 +1,115 @@
+/*
+ * Copyright 2010-2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ * You may not use this file except in compliance with the License.
+ * A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ * or in the "license" file accompanying this file. This file is distributed
+ * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+import {ConnectionConfig} from "./mqtt";
+import * as io from "./io";
+
+export class AwsIotMqttConnectionConfigBuilder {
+    private params: ConnectionConfig   
+    private tls_ctx_options?: io.TlsContextOptions
+
+    private constructor() {
+        this.params = {
+            client_id: '', 
+            host_name: '',
+            connect_timeout: 3000, 
+            port: io.is_alpn_available() ? 443 : 8883,
+            use_websocket: false,
+            clean_session: false,
+            keep_alive: undefined,
+            will: undefined,
+            username: '?SDK=NodeJSv2&Version=0.2.0',
+            password: undefined,
+            tls_ctx: undefined,
+        };
+    }
+
+    static new_mtls_builder_from_path(cert_path: string, key_path: string) {
+        let builder = new AwsIotMqttConnectionConfigBuilder();
+        builder.tls_ctx_options = io.TlsContextOptions.create_client_with_mtls(cert_path, key_path);
+        
+        if (io.is_alpn_available()) {
+            builder.tls_ctx_options.alpn_list = 'x-amzn-mqtt-ca';
+        }   
+        
+        return builder;
+    }  
+
+    with_certificate_authority_from_path(ca_path?: string, ca_file?: string) {
+        if (this.tls_ctx_options !== undefined) {
+            this.tls_ctx_options.override_default_trust_store(ca_path, ca_file);
+        }
+
+        return this;
+    }
+
+    with_endpoint(endpoint: string) {
+        this.params.host_name = endpoint;
+        return this;
+    }
+
+    with_client_id(client_id: string) {
+        this.params.client_id = client_id;
+        return this;
+    }
+
+    with_clean_session(clean_session: boolean) {
+        this.params.clean_session = clean_session;
+        return this;
+    }
+
+    with_use_websockets() {
+        this.params.use_websocket = true;
+
+        if (this.tls_ctx_options !== undefined) {
+            this.tls_ctx_options.alpn_list = undefined;
+            this.params.port = 443;
+        }
+
+        return this;
+    }
+
+    with_keep_alive_seconds(keep_alive: number) {
+        this.params.keep_alive = keep_alive;
+        return this;
+    }
+
+    with_timeout_ms(timeout_ms: number) {
+        this.params.timeout = timeout_ms;
+        return this;
+    }
+
+    with_will(will: string) {
+        this.params.will = will;
+        return this;
+    }
+
+    with_connect_timeout_ms(timeout: number) {
+        this.params.connect_timeout = timeout;
+        return this;
+    }
+
+    build() {
+        if (this.params.client_id === undefined || this.params.host_name === undefined) {
+            throw 'client_id and endpoint are required';
+        }
+
+        if (this.tls_ctx_options === undefined) {
+            throw 'tls options have to be specified'
+        }
+
+        this.params.tls_ctx = new io.ClientTlsContext(this.tls_ctx_options);       
+        return this.params;       
+    }
+}
